extended rem-sysapps, now checks if folder exists before attempting move

+++ windows_10/cleanup/rem-sysapps.ps1
@@ -8,6 +8,8 @@
     Script to Disable Pre-Installed SystemApps in Windows 10
     Cortana SearchUI.exe accesses the internet even with cortana disabled
     I remove the other sysapps as they are reporting/feedback apps
+    Note: Make sure Cortana is first in list if req'd as process respawns quickly
+    Also if you get access denied, run the script twice
     Requires Admin Access
     Twitter: @equilibriumuk 
 .NOTES
@@ -21,15 +23,17 @@
 
 $sysapppath = "$env:systemroot\SystemApps"
 $sysapps = @(
+    "Microsoft.Windows.Cortana_cw5n1h2txyewy"
     "ContactSupport_cw5n1h2txyewy"
     "ParentalControls_cw5n1h2txyewy"
     "WindowsFeedback_cw5n1h2txyewy"
-    "Microsoft.Windows.Cortana_cw5n1h2txyewy"
     )
 
 Write-Host ">> Killing Cortana Process"
 Get-Process *SearchUI* | Stop-Process -Force
 Write-Host ">> Moving Folders"
 foreach ($sysapp in $sysapps) {
-    mv $sysapppath\$sysapp $sysapppath\$sysapp"_disabled"
+    if (Test-Path $sysapppath\$sysapp) {
+        mv $sysapppath\$sysapp $sysapppath\$sysapp"_disabled"
+    }
 }
